restructure a bit
Restructure the files a bit to make more sense and easier to find
