+++ src/Faker.NET/Implementations/FakerInternet.cs
@@ -2,9 +2,9 @@ using System.Net;
 using Faker.NET.Common.Internet.Data;
 using Faker.NET.Extensions;
 using Faker.NET.Implementations;
-using Faker.NET.Interfaces;
 using Faker.NET.Interfaces.Definitions;
 using Faker.NET.Internet.Generators;
+using Faker.NET.Modules.Interfaces;
 
 namespace Faker.NET.Common.Internet;
