+++ src/Faker.NET/Implementations/FakerLocation.cs
@@ -1,9 +1,9 @@
 using Faker.NET.Common;
 using Faker.NET.Common.Location;
 using Faker.NET.Extensions;
-using Faker.NET.Interfaces;
 using Faker.NET.Interfaces.Definitions;
 using Faker.NET.Location;
+using Faker.NET.Modules.Interfaces;
 
 namespace Faker.NET.Implementations;
