+++ src/Faker.NET/Implementations/FakerPerson.cs
@@ -1,7 +1,7 @@
 using Faker.NET.Common.Person;
 using Faker.NET.Extensions;
-using Faker.NET.Interfaces;
 using Faker.NET.Interfaces.Definitions;
+using Faker.NET.Modules.Interfaces;
 
 namespace Faker.NET.Implementations;
