+++ src/Faker.NET/Implementations/FakerWord.cs
@@ -1,6 +1,6 @@
 using Faker.NET.Extensions;
-using Faker.NET.Interfaces;
 using Faker.NET.Interfaces.Definitions;
+using Faker.NET.Modules.Interfaces;
 
 namespace Faker.NET.Implementations;
