--- src/Faker.NET/Interfaces/IFakerPhoneNumber.cs
@@ -1,11 +1,0 @@
-﻿namespace Faker.NET.Interfaces
-{
-    public interface IFakerPhoneNumber
-    {
-        public string Number { get; }
-
-        public string NumberWithCountryCode { get; }
-
-        public int CountryCode { get; }
-    }
-}
