+++ src/Faker.NET/Interfaces/Modules/IFakerPhoneNumber.cs
@@ -1,0 +1,10 @@
+﻿namespace Faker.NET.Modules.Interfaces;
+
+public interface IFakerPhoneNumber
+{
+    public string Number { get; }
+
+    public string NumberWithCountryCode { get; }
+
+    public int CountryCode { get; }
+}
