+++ src/Faker.NET/Interfaces/Modules/IFakerWord.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace Faker.NET.Interfaces;
+namespace Faker.NET.Modules.Interfaces;
 
 public interface IFakerWord
 {
