+++ src/Faker.NET/Locales/EN/EnFakerPhoneNumber.cs
@@ -1,7 +1,7 @@
 using Faker.NET.Common;
 using Faker.NET.Extensions;
-using Faker.NET.Interfaces;
 using Faker.NET.Locales.EN.Data;
+using Faker.NET.Modules.Interfaces;
 
 namespace Faker.NET.Locales.EN;
